+++ src/bml.ts
@@ -5,8 +5,8 @@ import { analyze } from './analysis';
 
 // Wrap the main entrypoint function so we can attach further API parts to it
 export function entryFunc(bmlDocumentString: string,
-  renderSettings: RenderSettings | null, documentDir: string | null): string {
-  return render(bmlDocumentString, renderSettings, documentDir);
+  renderSettings: RenderSettings | null): string {
+  return render(bmlDocumentString, renderSettings);
 }
 
 entryFunc.analyze = analyze;